BUGFIX | Bump dependency
phpdocumentor/reflection-docblock has the option to install
a version of phpdocumentor/type-resolver below version 1.0

But the code now supports generic definitions in doc blocks
which are only supported in type-resolver version 1.0.0 and higher.

This commit also removes phpstan from the dev requirements as
it doesn't seem to be needed and moreover introduces class
ambiguity as it defines the same classes and namespaces as
refelction-docblock. And it is deprecated anyways.